move irmaseal.js
