+++ front-end/src/irmaseal.js
